chore: move DTOs&VOs
